refactor: centralize command abort session lookup
